+++ lib/mailkick/model.rb
@@ -1,21 +1,23 @@
 module Mailkick
   module Model
-    def has_subscriptions
+    def has_subscriptions(prefix: false)
+      prefix = prefix ? "mailkick_" : ""
+
       class_eval do
         has_many :mailkick_subscriptions, class_name: "Mailkick::Subscription", as: :subscriber
-        scope :subscribed, ->(list) { joins(:mailkick_subscriptions).where(mailkick_subscriptions: {list: list}) }
+        scope "#{prefix}subscribed", ->(list) { joins(:mailkick_subscriptions).where(mailkick_subscriptions: {list: list}) }
 
-        def subscribe(list)
+        define_method("#{prefix}subscribe") do |list|
           mailkick_subscriptions.where(list: list).first_or_create!
           nil
         end
 
-        def unsubscribe(list)
+        define_method("#{prefix}unsubscribe") do |list|
           mailkick_subscriptions.where(list: list).delete_all
           nil
         end
 
-        def subscribed?(list)
+        define_method("#{prefix}subscribed?") do |list|
           mailkick_subscriptions.where(list: list).exists?
         end
       end