+++ test/internal/app/models/admin.rb
@@ -1,0 +1,3 @@
+class Admin < ActiveRecord::Base
+  has_subscriptions prefix: true
+end