+++ test/internal/db/schema.rb
@@ -9,4 +9,8 @@ ActiveRecord::Schema.define do
   create_table :users do |t|
     t.string :email
   end
+
+  create_table :admins do |t|
+    t.string :email
+  end
 end